CI: fix broken ciimage builder script failing to correctly copy meson
Regression in commit 0af126fec798d6dbb0d1ad52168cc1f3f1758acd. We added
support for some "test cases/" stuff that actually relies on test files
being a symlink, but when testing the image builder, we copied the meson
repository contents into the docker container without telling python
that it is in fact super important to copy symlinks as symlinks.

As a result, the tests themselves ran fine when merging, but caused the
image builder to thereafter fail.